Add files via upload
Here are the Event classes, their fields deviate slightly from the class diagram to create them with repetition. 
set up in the source code as "(self, title, desc, date, dur, repeats = 0, freq = 2)" (2 is weekly)

The submitted status and the grade of a Class Event must be edited after creation, and are initialized by the constructor with the default values False and 0.00 respectively.